Initial support for HTML export
