+++ lib/hqmf-parser.rb
@@ -73,7 +73,10 @@ require_relative 'qrda-export/helper/aggregate_object_helper.rb'
 require_relative 'qrda-export/helper/code_system_helper.rb'
 require_relative 'qrda-export/helper/date_helper.rb'
 require_relative 'qrda-export/helper/view_helper.rb'
-require_relative 'qrda-export/helper/cat_1_view_helper.rb'
+require_relative 'qrda-export/helper/patient_view_helper.rb'
+require_relative 'qrda-export/helper/cat1_view_helper.rb'
+
+require_relative 'html-export/qdm-patient/qdm_patient.rb'
 
 require_relative 'qrda-export/catI-r5/qrda1_r5.rb'
 require_relative 'qrda-export/catIII-r2-1/qrda3_r21.rb'
