+++ lib/html-export/qdm-patient/_header_css.mustache
@@ -1,0 +1,61 @@
+<style type="text/css">
+  body {
+    color: #003366;
+    background-color: #FFFFFF;
+    font-family: Verdana, Tahoma, sans-serif;
+    font-size: 11px;
+  }
+  .criteria-heading {
+    font-weight: bold;
+  }
+  .div-table{
+    display: table;
+    width: 100%;
+  }
+  .div-table-body {
+    display: table-row-group;
+  }
+  .div-table-row {
+    display: table-row;
+  }
+  .div-head-row {
+    display: table-row;
+  }
+  .div-table-cell, .div-table-head {
+    border: 1px solid #999999;
+    display: table-cell;
+    padding: 3px 10px;
+  }
+  .div-table-head--no-border {
+    display: table-cell;
+  }
+  .div-table-heading {
+    background-color: #EEE;
+    display: table-header-group;
+    font-weight: bold;
+  }
+  .patient_narr_tr {
+    background-color: #ccccff;
+  }
+  .patient_narr_th {
+    background-color: #3399ff;
+  }
+  .td_label{
+    font-weight: bold;
+    color: white;
+  }
+  .narr_tr {
+    background-color: #ffffcc;
+  }
+  .narr_th-wide {
+    background-color: #ffd700;
+    width: 28%;
+  }
+  .narr_th-narrow {
+    background-color: #ffd700;
+    width: 16%;
+  }
+  .div-table-cell-label {
+    display: none;
+  }
+</style>
