+++ lib/html-export/qdm-patient/data_element/_data_element.mustache
@@ -1,0 +1,4 @@
+<div class="div-table-cell"><span class="description-heading">{{description}}</span></div>
+<div class="div-table-cell"><div class="div-table-cell-label">Codes: </div>{{> data_element/_data_element_codes}}</div>
+<div class="div-table-cell"><div class="div-table-cell-label">Time Elements: </div>{{> data_element/_data_element_times}}</div>
+<div class="div-table-cell"><div class="div-table-cell-label">Other Fields: </div>{{> data_element/_data_element_other_fields}}</div>
