+++ lib/html-export/qdm-patient/data_element/_data_element_codes.mustache
@@ -1,0 +1,9 @@
+<div class="div-table">
+<div class="div-table-body">
+{{#dataElementCodes}}
+<div class="div-head-row">
+	<div class="div-table-head--no-border"><span class="criteria-heading">{{codeSystem}}:</span>  {{code}}</div>
+</div>
+{{/dataElementCodes}}
+</div>
+</div>
