+++ lib/html-export/qdm-patient/data_element/_data_element_other_fields.mustache
@@ -1,0 +1,106 @@
+<div class="div-table">
+<div class="div-table-body">
+{{#admissionSource}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Admission Source:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/admissionSource}}
+{{#anatomicalApproachSite}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Anatomical Approach Site:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/anatomicalApproachSite}}
+{{#anatomicalLocationSite}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Anatomical Location Site:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/anatomicalLocationSite}}
+{{#components}}
+	{{#code}}	
+		<div class="div-head-row">
+			<div class="div-table-head--no-border"><span class="criteria-heading">Component Code:</span>  {{{code_code_system_string}}}</div>
+		</div>
+	{{/code}}
+	{{#result}}
+		<div class="div-head-row">
+			<div class="div-table-head--no-border"><span class="criteria-heading">Component Result:</span>  {{{code_code_system_string}}}</div>
+		</div>
+	{{/result}}
+{{/components}}
+{{#diagnoses}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Diagnosis:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/diagnoses}}
+{{#dischargeDisposition}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Discharge Disposition:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/dischargeDisposition}}
+{{#facilityLocation}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Facility Location:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/facilityLocation}}
+{{#facilityLocations}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Facility:</span>  {{{facility_string}}}</div>
+	</div>
+{{/facilityLocations}}
+{{#frequency}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Frequency:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/frequency}}
+{{#lengthOfStay}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Length Of Stay:</span>  {{{unit_string}}}</div>
+	</div>
+{{/lengthOfStay}}
+{{#negationRationale}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Negation Reason:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/negationRationale}}
+{{#ordinality}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Ordinality:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/ordinality}}
+{{#principalDiagnosis}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Principal Diagnosis:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/principalDiagnosis}}
+{{#reason}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Reason:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/reason}}
+{{#route}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Route:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/route}}
+{{#severity}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Severity:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/severity}}
+{{#status}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Status:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/status}}
+{{#type}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Type:</span>  {{{code_code_system_string}}}</div>
+	</div>
+{{/type}}
+{{#result}}
+	<div class="div-head-row">
+		<div class="div-table-head--no-border"><span class="criteria-heading">Result:</span>  {{{result_string}}}</div>
+	</div>
+{{/result}}
+</div>
+</div>
