+++ lib/qrda-export/catI-r5/qrda1_r5.rb
@@ -3,6 +3,7 @@ class Qrda1R5 < Mustache
   include Qrda::Export::Helper::DateHelper
   include Qrda::Export::Helper::ViewHelper
   include Qrda::Export::Helper::Cat1ViewHelper
+  include Qrda::Export::Helper::PatientViewHelper
 
   self.template_path = __dir__
